requirements.txt
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
langchain>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
coverage>=7.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
datasets>=2.14.0
httpx>=0.25.0
aiofiles>=23.2.0
redis>=5.0.0
aiohttp>=3.9.0
boto3>=1.34.0
slowapi>=0.1.9
asyncpg>=0.29.0
alembic>=1.13.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
faker>=20.0.0
click>=8.1.0
tabulate>=0.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
aiocache>=0.12.2

# Performance Optimization Packages
msgpack>=1.0.7          # Fast serialization
ujson>=5.9.0            # Fast JSON
orjson>=3.9.10          # Fastest JSON
hiredis>=2.2.3          # C parser for Redis
aioredis>=2.0.1         # Async Redis
aiosqlite>=0.19.0       # Async SQLite
prometheus-client>=0.19.0  # Metrics
py-spy>=0.3.14          # Profiling
memory-profiler>=0.61.0 # Memory profiling